html 렌더링 기능과 api 구분
