+++ src/main/java/gift/controller/WishlistController.java
@@ -41,27 +41,6 @@ public class WishlistController {
     }
 
 
-    @GetMapping
-    public ResponseEntity<Map<String, Object>> getWishlistItems(
-            @RequestParam("email") String email,
-            Pageable pageable) {
-        Page<Product> productPage = wishlistService.getWishlistByEmail(email, pageable);
-        Map<String, Object> response = new HashMap<>();
-        List<ProductDto> productDtoList = new ArrayList<>();
-
-        for (Product product : productPage.getContent()) {
-            ProductDto dto = new ProductDto(product);
-            productDtoList.add(dto);
-        }
-
-        response.put("content", productDtoList);
-        response.put("currentPage", productPage.getNumber() + 1);
-        response.put("totalPages", productPage.getTotalPages());
-        response.put("hasNext", productPage.hasNext());
-        response.put("hasPrevious", productPage.hasPrevious());
-        return ResponseEntity.ok(response);
-    }
-
     @DeleteMapping("/{productId}")
     public ResponseEntity<Void> deleteWishlistItem(@RequestHeader("Authorization") String token, @PathVariable Long productId) {
         String email = jwtTokenProvider.getEmail(token.substring(7));
