+++ src/main/resources/templates/add-product.html
@@ -46,9 +46,8 @@
         <!-- 옵션 필드가 추가될 위치 -->
     </div>
     <button type="button" onclick="addOptionField()">Add Option</button><br/>
-
-
     <button type="submit">Add Product</button>
+    <a href="/view/products">Cancel</a>
 </form>
 </body>
 </html>
