+++ src/main/resources/templates/error/400.html
@@ -9,6 +9,6 @@
 <h1>400 Bad Request</h1>
 <p th:text="${error}"></p>
 400 Error입니다.<br>
-<a href="/products">Back to Product List</a>
+<a href="/view/products">Back to Product List</a>
 </body>
 </html>
